pyproject.toml
[project]
name = "feedlyai-work"
version = "0.1.0"
description = "YH's part of FeedlyAI - YOLO/Planner/Overlay/Eval/Judge"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",
    "python-multipart==0.0.9",
    "pydantic==2.7.4",
    "Pillow==10.3.0",
    "requests==2.32.3",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "prometheus-client==0.19.0",
    # LLaVa 모델 관련
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",  # 8-bit 양자화 지원 (선택사항, 메모리 절약)
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

